Split SvgDocument
